tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/size.attributes.style.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: size.attributes.style</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>size.attributes.style</h1>
<p class="desc">Canvas size is independent of CSS resizing</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="50" height="30" style="width: 100px; height: 50px"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.style.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Canvas size is independent of CSS resizing");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 50, "canvas.width", "50");
_assertSame(canvas.height, 30, "canvas.height", "30");


});
</script>
